Implemented first version of Shared Prefs affecting randomisation process implemented tabs on perk preferences page. Updated pop return value from buildConfiguration.dart
